Moved CharacterGallery related signal bus events to ModelController.cs
